'makefile medio compila y tal'

+++ libft/ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: sgomez-p <sgomez-p@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/16 12:51:24 by sgomez-p          #+#    #+#             */
-/*   Updated: 2022/09/26 12:55:47 by sgomez-p         ###   ########.fr       */
+/*   Updated: 2023/03/01 12:59:29 by sgomez-p         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i] != '\0')
